+++ src/FeuThingAction.h
@@ -18,9 +18,9 @@ private:
     FeuCalculable *mWhat; // From the "what" attribute
 
 public:
-    float proc(FeuThing *contextThing = NULL); // Run the action handler
-    virtual float getAttributeValue(std::string);
-    virtual void  setAttributeValue(std::string,float);
+    double proc(FeuThing *contextThing = NULL); // Run the action handler
+    virtual double getAttributeValue(std::string);
+    virtual void  setAttributeValue(std::string,double);
 
 };
