+++ src/FeuThingClass.h
@@ -21,8 +21,8 @@ public:
 public:
     void harvestPic(FeuThingPic *pic);
     void renderPic(FeuThingPic *pic);
-    virtual float getAttributeValue(std::string);
-    virtual void  setAttributeValue(std::string,float);
+    virtual double getAttributeValue(std::string);
+    virtual void  setAttributeValue(std::string,double);
 };
